requirements.txt
streamlit>=1.35.0
streamlit-authenticator==0.3.3
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
openpyxl>=3.1.2
pandas>=2.2.0
python-pptx>=0.6.23
PyYAML>=6.0
bcrypt>=4.1.0
